Automatic line ending normalization
In github for windows, went to settings for the mcex project and saw
these messages:

"The .gitattributes file controls line ending normalization and helps
Git decide which files are binary and which are text.

*Add the recommended .getattributes file.*

This repository's line endings aren't normalized.
The best solution is to add the recommended file and make a one-time
commit to fix line endings. If you don't want to use line ending
normalization you can *turn this feature off.*"

Clicking the *Add the recommended ...* link caused these changes.